Pull CollapsibleContentView+Style into separate file
